Splitting out the entities to a separate directory
