+++ orders_api/src/main.rs
@@ -1,142 +1,18 @@
 use actix_web::error::HttpError;
 use actix_web::web::Data;
-use actix_web::{guard, web, App, HttpResponse, HttpServer, ResponseError};
+use actix_web::{guard, web, App, HttpResponse, HttpServer};
 use async_graphql::{
     http::{playground_source, GraphQLPlaygroundConfig},
-    Context, EmptySubscription, Enum, FieldResult, Object, Schema, SimpleObject,
+    EmptySubscription, Schema,
 };
 use async_graphql_actix_web::{GraphQLRequest, GraphQLResponse};
 
-use derive_more::{Display, From};
+use deadpool_postgres::{Client, Config, Pool};
+use tokio_postgres::NoTls;
 
-use tokio_pg_mapper::Error as PGMError;
-use tokio_postgres::error::Error as PGError;
-use tokio_postgres::{NoTls};
-
-use deadpool_postgres::{Client, Config, Pool, PoolError};
-
-use orders_api::models::Burger;
+use orders_api::models::{QueryRoot, CustomError, MutationRoot};
 
 pub type ShopSchema = Schema<QueryRoot, MutationRoot, EmptySubscription>;
-
-pub struct MutationRoot;
-
-#[Object]
-impl MutationRoot {
-    async fn add(&self, a: i32, b: i32) -> i32 {
-        a + b
-    }
-}
-
-#[derive(Enum, Eq, PartialEq, Copy, Clone)]
-pub enum Size {
-    Small,
-    Medium,
-    Large,
-}
-
-#[derive(SimpleObject, Clone)]
-pub struct Fries {
-    id: Option<String>,
-    size: Size,
-    cost: i32,
-}
-
-#[derive(Enum, PartialEq, Eq, Clone, Copy)]
-pub enum DrinkType {
-    Water,
-    Coke,
-    Sprite,
-    OrangeJuice,
-}
-
-#[derive(SimpleObject, Clone)]
-pub struct Drink {
-    id: Option<String>,
-    size: Size,
-    drink_type: DrinkType,
-    cost: i32,
-}
-
-#[derive(SimpleObject, Clone)]
-pub struct Meal {
-    id: Option<String>,
-    name: String,
-    burger: Burger,
-    fries: Fries,
-    drink: Drink,
-    cost: i32,
-}
-
-// this is my schema?
-#[derive(SimpleObject, Clone)]
-pub struct Menu {
-    hello: String,
-    meals: Vec<Meal>,
-    burgers: Vec<Burger>,
-}
-
-impl Menu {
-    pub async fn new() -> Self {
-        Self {
-            hello: "hello".to_string(),
-            meals: [Meal {
-                id: Some("123".to_string()),
-                name: "Standard meal".to_string(),
-                cost: 1200,
-                burger: Burger {
-                    id: Some("124".to_string()),
-                    burger_type: "Cheesy".to_string(),
-                    cost: 500,
-                },
-                fries: Fries {
-                    id: Some("125".to_string()),
-                    size: Size::Large,
-                    cost: 300,
-                },
-                drink: Drink {
-                    id: Some("126".to_string()),
-                    size: Size::Large,
-                    drink_type: DrinkType::Water,
-                    cost: 200,
-                },
-            }]
-            .to_vec(),
-            burgers: [Burger {
-                id: Some("124".to_string()),
-                burger_type: "Cheesy".to_string(),
-                cost: 500,
-            }]
-            .to_vec(),
-        }
-    }
-}
-
-pub struct QueryRoot;
-
-// this is the implementation of the query
-#[Object]
-impl QueryRoot {
-    async fn hello(&self) -> FieldResult<&str> {
-        Ok("hello")
-    }
-    async fn menu(&self) -> FieldResult<Menu> {
-        Ok(Menu::new().await)
-    }
-    async fn get_burgers(&self, ctx: &Context<'_>) -> FieldResult<Vec<Burger>> {
-        let db = &ctx
-            .data_unchecked::<Pool>()
-            .get()
-            .await
-            .map_err(CustomError::PoolError)?;
-
-        let result = db.query("select * from burgers", &[])
-            .await
-            .unwrap();
-
-        Ok(result.into_iter().map(|row| Burger::from(row)).collect::<Vec<Burger>>().to_vec())
-    }
-}
 
 async fn index(schema: Data<ShopSchema>, req: GraphQLRequest) -> GraphQLResponse {
     schema.execute(req.into_inner()).await.into()
@@ -147,28 +23,6 @@ async fn index_playground() -> Result<HttpResponse, HttpError> {
     Ok(HttpResponse::Ok()
         .content_type("text/html; charset=utf-8")
         .body(source))
-}
-
-#[derive(Display, From, Debug)]
-pub enum CustomError {
-    NotFound,
-    PGError(PGError),
-    PGMError(PGMError),
-    PoolError(PoolError),
-}
-
-impl std::error::Error for CustomError {}
-
-impl ResponseError for CustomError {
-    fn error_response(&self) -> HttpResponse {
-        match *self {
-            CustomError::NotFound => HttpResponse::NotFound().finish(),
-            CustomError::PoolError(ref err) => {
-                HttpResponse::InternalServerError().body(err.to_string())
-            }
-            _ => HttpResponse::InternalServerError().finish(),
-        }
-    }
 }
 
 async fn initialise_db(db_pool: web::Data<Pool>) -> Result<HttpResponse, CustomError> {
