+++ orders_api/src/models/fry.rs
@@ -1,0 +1,9 @@
+use crate::models::size::Size;
+use async_graphql::SimpleObject;
+
+#[derive(SimpleObject, Clone)]
+pub struct Fries {
+    pub id: Option<String>,
+    pub size: Size,
+    pub cost: i32,
+}
