+++ orders_api/src/models/mutation.rs
@@ -1,0 +1,9 @@
+use async_graphql::Object;
+pub struct MutationRoot;
+
+#[Object]
+impl MutationRoot {
+    async fn add(&self, a: i32, b: i32) -> i32 {
+        a + b
+    }
+}
