+++ orders_api/src/models/size.rs
@@ -1,0 +1,8 @@
+use async_graphql::Enum;
+
+#[derive(Enum, Eq, PartialEq, Copy, Clone)]
+pub enum Size {
+    Small,
+    Medium,
+    Large,
+}
